+++ math.cs
@@ -4,27 +4,27 @@ using System;
 namespace polygot
 {
     
-     public class Matrix: List<List<float>> {  }
+     public class Matrix: List<List<double>> {  }
    public class math
     {
         public void zeroes(Matrix M,int n){
             for(int i=0;i<n;i++){
-                List<float> row = new List<float>(new float[n]);
-               // List<float> row(n,0.0);
+                List<double> row = new List<double>(new double[n]);
+               // List<double> row(n,0.0);
                 M.Add(row);
             }
         }
 
         public void zeroes(Matrix M,int n,int m){
             for(int i=0;i<n;i++){
-               List<float> row = new List<float>(new float[m]);
+               List<double> row = new List<double>(new double[m]);
                 M.Add(row);
             }
         }
 
-       public void zeroes(List<float> v,int n){
+       public void zeroes(List<double> v,int n){
             for(int i=0;i<n;i++){
-                v.Add((float)0.0);
+                v.Add((double)0.0);
             }
         }
 
@@ -35,8 +35,8 @@ namespace polygot
                     copy[i][j] = A[i][j];
         }
 
-     public   float calculateMember(int i,int j,int r,Matrix A,Matrix B){
-            float member = 0;
+     public   double calculateMember(int i,int j,int r,Matrix A,Matrix B){
+            double member = 0;
             for(int k=0;k<r;k++)
                 member += A[i][k]*B[k][j];
             return member;
@@ -53,9 +53,9 @@ namespace polygot
             return R;
         }
 
-   public     void productMatrixVector(Matrix A, List<float>  v, List<float>  R){
+   public     void productMatrixVector(Matrix A, List<double>  v, List<double>  R){
             for(int f=0;f<A.Count;f++){
-                float cell = 0.0F;
+                double cell = 0.0F;
                 for(int c=0;c<v.Count;c++){
                     cell += A[f][c]*v[c];
                 }
@@ -63,7 +63,7 @@ namespace polygot
             }
         }
 
-     public   void productRealMatrix(float real,Matrix M,Matrix R){
+     public   void productRealMatrix(double real,Matrix M,Matrix R){
             zeroes(R,M.Count);
             for(int i=0;i<M.Count;i++)
                 for(int j=0;j<M[0].Count;j++)
@@ -79,16 +79,16 @@ namespace polygot
                 M[i].Remove(M[i][0+j]);
         }
 
-   public     float determinant(Matrix M){
+   public     double determinant(Matrix M){
             if(M.Count == 1) return M[0][0];
             else{
-                float det=0.0F;
+                double det=0.0F;
                 for(int i=0;i<M[0].Count;i++){
                     Matrix minor = new Matrix();
                     copyMatrix(M,minor);
                     getMinor(minor,0,i);
                     
-                    det += (float) Math.Pow(-1,i)*M[0][i]*determinant(minor);
+                    det += (double) Math.Pow(-1,i)*M[0][i]*determinant(minor);
                 }
                 return det;
             }
@@ -102,7 +102,7 @@ namespace polygot
                     Matrix minor = new Matrix();
                     copyMatrix(M,minor);
                     getMinor(minor,i,j);
-                    Cof[i][j] = (float)Math.Pow(-1,i+j)*determinant(minor);
+                    Cof[i][j] = (double)Math.Pow(-1,i+j)*determinant(minor);
                 }
             }
         }
@@ -119,7 +119,7 @@ namespace polygot
             Matrix Cof =  new Matrix() ;
             Matrix Adj = new Matrix();
             Console.WriteLine("Calculo de determinante...\n");
-            float det = determinant(M);
+            double det = determinant(M);
             if(det == 0) System.Environment.Exit(1);
             Console.WriteLine("Iniciando calculo de cofactores...\n");
             cofactors(M,Cof);
